+++ src/datasources/data-frame/components/v2/DataFrameVariableQueryEditorV2.tsx
@@ -2,8 +2,9 @@ import { Combobox, ComboboxOption } from "@grafana/ui";
 import { InlineField } from "core/components/InlineField";
 import { FloatingError } from "core/errors";
 import { INLINE_LABEL_WIDTH } from "datasources/data-frame/constants/v2/DataFrameQueryEditorV2.constants";
-import { DataFrameVariableQuery, DataFrameVariableQueryType, Props } from "datasources/data-frame/types";
+import { DataFrameVariableQueryType, Props } from "datasources/data-frame/types";
 import React from "react";
+import { DataFrameQueryBuilderWrapper } from "./query-builders/DataFrameQueryBuilderWrapper";
 
 export const DataFrameVariableQueryEditorV2: React.FC<Props> = ({ query, onChange, datasource }: Props) => {
     const migratedQuery = datasource.processVariableQuery(query);
@@ -13,7 +14,14 @@ export const DataFrameVariableQueryEditorV2: React.FC<Props> = ({ query, onChang
     ];
 
     const onQueryTypeChange = (option: ComboboxOption<string>) => {
-        onChange({ ...migratedQuery, queryType: option.value } as DataFrameVariableQuery);
+        onChange({ ...migratedQuery, queryType: option.value as DataFrameVariableQueryType });
+    };
+
+    const onDataTableFilterChange = (event?: Event | React.FormEvent<Element>) => {
+        if (event) {
+            const dataTableFilter = (event as CustomEvent).detail.linq;
+            onChange({ ...migratedQuery, dataTableFilter });
+        }
     };
 
     return (
@@ -30,7 +38,11 @@ export const DataFrameVariableQueryEditorV2: React.FC<Props> = ({ query, onChang
                     placeholder={placeholder}
                     width={40} />
             </InlineField>
-            {/* TODO: #3463943 - Integrate query builder wrapper to variable query editor */}
+            <DataFrameQueryBuilderWrapper
+                datasource={datasource}
+                dataTableFilter={migratedQuery.dataTableFilter}
+                onDataTableFilterChange={onDataTableFilterChange}
+            />
             <FloatingError
                 message={datasource.errorTitle}
                 innerMessage={datasource.errorDescription}